# specs/shared/openapi.yaml
openapi: 3.1.0
info:
  title: Order-To-Cash — Gateway REST API
  version: 1.0.0
  summary: The single external entry point of the Order-To-Cash system.
  description: |
    **Stack-agnostic** REST contract of the Gateway / BFF, reused **verbatim**
    by assessments **#7**, **#8** and **#9**. It is the only surface the outside
    world may touch: the web application, the API tests, the end-to-end tests
    and the demo automation workflows all speak this contract and nothing else.
    Nothing outside the system connects to a database, to the fact stream or to
    the RPC transport.

    Companion documents: `domain-model.md`, `saga.md`, `requirements.md`
    (`R1`–`R61`) and `asyncapi.yaml` — the messaging contract this gateway
    translates into.

    ## What the Gateway is, and is not

    - It **owns no invariant and no aggregate.** It translates HTTP into
      request-reply commands and queries over the RPC transport, and serves
      list and detail queries from the **read model** (R54).
    - **Command endpoints** (`POST /orders`, `POST /orders/{id}/cancel`,
      `POST /invoices/{id}/payments`, `POST /stock/replenish`) are RPC calls to
      the owning context. A 2xx here means *the command was accepted*, never
      *the saga finished*.
    - **Query endpoints** (`GET /orders`, `GET /orders/{id}`) read the
      projected read model **only** — never a write model, never a
      cross-context join.
...
      projector's update signal (R55).

...

...
    instantly queryable.

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: Issue a bearer token.
      description: |
        Exchanges operator credentials for a bearer token. A single operator
        identity is assumed — the model is deliberately not multi-tenant
        (`domain-model.md` §9).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Bad credentials.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentUser
      summary: Describe the authenticated operator.
      responses:
        '200':
          description: The current operator.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /orders/stream:
    get:
      tags: [stream]
      operationId: streamOrderEvents
      summary: Live saga progress as server-sent events.
      description: |
        A **server-sent event stream** (`text/event-stream`), fed by the
        projector's read-model update signal (R55). One-directional by design:
        the client never sends anything back on it, which is why an SSE stream
        rather than a bidirectional socket.

        ### Frame format

        Each frame is a standard SSE block:

        ```
        id: 1755511234567-17
        event: order.updated
        data: {"orderId":"…","status":"confirmed", …}

        ```

        - `event` names the event type. The types emitted are listed below.
        - `data` is a single-line JSON document whose schema depends on `event`.
        - `id` is an opaque, monotonically increasing cursor **on `order.updated`
          and `timeline.appended` only** — the two event types the replay buffer
          stores. **`ping` deliberately carries no `id` line.**

          This is not an oversight: a browser's `EventSource` updates its
          internal `lastEventId` from *any* dispatched event that carries one,
          heartbeat or not. `ping` fires on a fixed interval and real content
          fires rarely, so if `ping` carried an id, the client's remembered
          resume point would almost always be a heartbeat's — and a heartbeat's
          id is never in the replay buffer (there is nothing to replay), so
          `Last-Event-ID` would name it and reconnection would answer
          `resumed: false` on nearly every disconnect, forcing a re-fetch even
          when nothing was actually missed. Omitting `id` from `ping` keeps a
          client's remembered cursor pointed at real content, so reconnection
          `resumed: false` means what §2 below says it means: the buffer
          genuinely could not find that cursor, not "you happened to disconnect
          between two heartbeats." `ping` still carries its documented `event`
          and `data` (`StreamPing`) — only the `id` line is absent. Found by
          hand-testing after feature 26 landed; `stream.ready`'s cursor has the
          identical property (minted, never stored) for the same reason, but is
          sent only once per connection so the interaction is not observable in
          practice the way `ping`'s is.

        ### Event types

        | `event` | `data` schema | Carries `id`? | Meaning |
        |---|---|---|---|
        | `order.updated` | `OrderStreamUpdate` | Yes — resumable | A read-model document changed — status, references or totals |
        | `timeline.appended` | `TimelineStreamEntry` | Yes — resumable | One fact was appended to an order's timeline |
        | `stream.ready` | `StreamReady` | No | Sent once on connect, carrying the cursor the stream resumed from |
        | `ping` | `StreamPing` | No | Keep-alive, sent on an idle interval so proxies do not close the connection |

        ### Reconnection

        On reconnect the client sends the id of the last frame it processed in
        the `Last-Event-ID` request header (browsers do this automatically), and
        the stream **resumes after that cursor** from a bounded replay buffer.
        Two honest limitations, stated rather than hidden:

        1. The buffer is bounded. If the client was away longer than the buffer
           holds, the stream replies `stream.ready` with `resumed: false` and
           the client must re-fetch `GET /orders/{id}` (or `GET /orders`) to
           resynchronise. The stream is a *notification* channel; the read model
           is the source of truth.
        2. Delivery is at-least-once. A frame may repeat after a reconnect, so
           clients deduplicate on the `eventId` inside `data` — the same
           discipline every fact consumer applies (R51).
      parameters:
        - name: orderId
          in: query
          description: Restrict the stream to one order. Omit to receive updates for every order.
          schema:
            $ref: '#/components/schemas/UniqueId'
        - name: Last-Event-ID
          in: header
          description: Cursor of the last frame the client processed. Sent automatically by browser EventSource implementations on reconnect.
          schema:
            type: string
      responses:
        '200':
          description: |
            The event stream. It stays open until the client disconnects; there
            is no terminal frame.
          headers:
            Cache-Control:
              schema:
                type: string
                const: no-cache
            Connection:
              schema:
                type: string
                const: keep-alive
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  A never-ending sequence of SSE frames. OpenAPI cannot model a
                  stream's frames, so the payload is typed as the raw media
                  type and the per-event `data` schemas are given in
                  `x-sse-events` below and in `components.schemas`.
              examples:
                progression:
                  summary: Two frames of a saga in progress
                  value: |
                    id: 1755511234567-17
                    event: order.updated
                    data: {"eventId":"5b7a0f1e-1f22-4c1f-9f2a-8c1a1f2b3c4d","orderId":"9f1e2d3c-4b5a-4c6d-8e7f-0a1b2c3d4e5f","orderReference":"ORD-000042","status":"stock_reserved","occurredAt":"2026-08-18T10:15:02.100Z"}

                    id: 1755511234901-18
                    event: timeline.appended
                    data: {"eventId":"7d8e9f0a-1b2c-4d3e-8f4a-5b6c7d8e9f01","orderId":"9f1e2d3c-4b5a-4c6d-8e7f-0a1b2c3d4e5f","eventType":"credit.rejected.v1","occurredAt":"2026-08-18T10:15:04.120Z","summary":"Credit rejected (simulated_cents_rule)"}
        '401':
          $ref: '#/components/responses/Unauthorized'
      x-sse-events:
        order.updated:
          schema:
            $ref: '#/components/schemas/OrderStreamUpdate'
        timeline.appended:
          schema:
            $ref: '#/components/schemas/TimelineStreamEntry'
        stream.ready:
          schema:
            $ref: '#/components/schemas/StreamReady'
        ping:
          schema:
            $ref: '#/components/schemas/StreamPing'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Bearer token obtained from `POST /auth/login`. Required on every
        endpoint except `POST /auth/login`, `GET /health/live`,
        `GET /health/ready` and `GET /docs`.

        The demo automation workflows authenticate exactly like any other
        client: they log in, then send the token. They hold no database
        credential and no broker credential, because they never talk to either.
...
  responses:
    BadRequest:
      description: The request was malformed or failed schema validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ValidationProblem'
    Unauthorized:
      description: Missing, expired or invalid bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: |
        The client exceeded the rate limit of the endpoint it called. The
        request was **not** applied and no fact was emitted; the client may
        retry once the interval named by `Retry-After` has elapsed.
      headers:
        Retry-After:
          description: |
            How long the client must wait before retrying, as a whole number of
            **seconds**. Delta-seconds only: the HTTP-date form RFC 9110 also
            permits is deliberately **not** used by this API, so a client may
            parse this header as an integer without branching on its form. The
            value is the remaining lifetime of the limiter's current window;
            the window length itself is per-assessment configuration and is
            deliberately not fixed by this contract.
          required: true
          schema:
            type: integer
            minimum: 0
            examples:
              - 60
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
...
    UniqueId:
      type: string
      format: uuid
      description: An opaque, globally unique identifier generated inside the domain, never by a store.
      examples:
        - 9f1e2d3c-4b5a-4c6d-8e7f-0a1b2c3d4e5f
    Instant:
      type: string
      format: date-time
      description: A UTC instant, ISO-8601.
      examples:
        - '2026-08-18T10:15:00.000Z'
...
    MinorUnits:
      type: integer
      format: int64
      description: |
        An amount as a whole count of the currency's smallest denomination.
        **Never** a decimal, float or major-unit value (R1) — `€1,242.50` is
        `124250`. Used where the enclosing object already declares its
        `currency`; where an amount travels alone, use `Money`.
      examples:
        - 124250
...
    OrderReference:
      type: string
      pattern: '^ORD-[0-9]{6,}$'
      examples:
        - ORD-000042
    DespatchReference:
      type: string
      pattern: '^DES-[0-9]{6,}$'
      examples:
        - DES-000031
    InvoiceReference:
      type: string
      pattern: '^INV-[0-9]{6,}$'
      examples:
        - INV-000027
...
    PaymentReference:
      type: string
      minLength: 1
      maxLength: 30
      description: The remittance's own reference — **the idempotency key** of `POST /invoices/{id}/payments` (R48).
      examples:
        - PAY-2026-08-18-000019
    OrderStatus:
      type: string
      enum:
        - placed
        - stock_reserved
        - credit_approved
        - confirmed
        - despatched
        - invoiced
        - paid
        - completed
        - cancelled
      description: The order state machine, which is also the saga state. `completed` and `cancelled` are terminal.
    CancellationReason:
      type: string
      enum:
        - stock_rejected
        - credit_rejected
        - operator_cancelled
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required: [username, password]
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          const: Bearer
        expiresIn:
          type: integer
          minimum: 1
          description: Lifetime of the token in seconds.
      required: [accessToken, tokenType, expiresIn]
    CurrentUser:
      type: object
      properties:
        username:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
      required: [username, roles]

    # ─────────────────────────── orders — commands ──────────────────────────
...
    OrderTotals:
      type: object
      description: All three values are minor units in the order's `currency`, derived by the aggregate.
      properties:
        initialAmount:
          $ref: '#/components/schemas/MinorUnits'
        initialDiscount:
          $ref: '#/components/schemas/MinorUnits'
        totalAmount:
          $ref: '#/components/schemas/MinorUnits'
      required: [initialAmount, initialDiscount, totalAmount]
...
    OrderReferences:
      type: object
      description: Business references picked up as the saga progressed. Null until the corresponding fact has been projected.
      properties:
        despatchReference:
          oneOf:
            - $ref: '#/components/schemas/DespatchReference'
            - type: 'null'
        invoiceReference:
          oneOf:
            - $ref: '#/components/schemas/InvoiceReference'
            - type: 'null'
        paymentReference:
          oneOf:
            - $ref: '#/components/schemas/PaymentReference'
            - type: 'null'
...
    OrderStreamUpdate:
      type: object
      description: '`data` of an `order.updated` frame — a read-model document changed.'
      properties:
        eventId:
          allOf:
            - $ref: '#/components/schemas/UniqueId'
          description: The fact that caused the update. Clients deduplicate on it, because stream delivery is at-least-once.
        orderId:
          $ref: '#/components/schemas/UniqueId'
        orderReference:
          $ref: '#/components/schemas/OrderReference'
        status:
          $ref: '#/components/schemas/OrderStatus'
        cancellationReason:
          oneOf:
            - $ref: '#/components/schemas/CancellationReason'
            - type: 'null'
        references:
          $ref: '#/components/schemas/OrderReferences'
        totals:
          $ref: '#/components/schemas/OrderTotals'
        occurredAt:
          $ref: '#/components/schemas/Instant'
      required: [eventId, orderId, status, occurredAt]
    TimelineStreamEntry:
      type: object
      description: '`data` of a `timeline.appended` frame — one fact was appended to an order timeline.'
      properties:
        eventId:
          $ref: '#/components/schemas/UniqueId'
        causationId:
          allOf:
            - $ref: '#/components/schemas/UniqueId'
          description: Same field as `TimelineEntry.causationId`, carried onto the stream frame.
        orderId:
          $ref: '#/components/schemas/UniqueId'
        orderReference:
          $ref: '#/components/schemas/OrderReference'
        eventType:
          type: string
        occurredAt:
          $ref: '#/components/schemas/Instant'
        summary:
          type: string
      required: [eventId, orderId, eventType, occurredAt, summary]
    StreamReady:
      type: object
      description: '`data` of the `stream.ready` frame sent once on connect.'
      properties:
        cursor:
          type: string
          description: The cursor the stream is resuming from.
        resumed:
          type: boolean
          description: |
            True when a `Last-Event-ID` was honoured from the replay buffer.
            **False** when it was too old — the client must re-fetch the
            affected orders to resynchronise; the stream is a notification
            channel, the read model is the source of truth.
        orderId:
          oneOf:
            - $ref: '#/components/schemas/UniqueId'
            - type: 'null'
          description: Present when the subscription is restricted to one order.
      required: [cursor, resumed]
    StreamPing:
      type: object
      description: '`data` of a `ping` keep-alive frame.'
      properties:
        at:
          $ref: '#/components/schemas/Instant'
      required: [at]

    # ───────────────────────────── fulfillment ──────────────────────────────
...
    Problem:
      type: object
      description: |
        RFC 9457 problem document. `code` is stable and machine-readable and is
        what tests assert on; `title` and `detail` are for humans and must never
        be parsed.
      properties:
        type:
          type: string
          format: uri
          default: about:blank
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
          description: Stable machine-readable error code.
          examples:
            - ORDER_NOT_CANCELLABLE
            - INVOICE_ALREADY_PAID
            - PAYMENT_MISMATCH
            - PAYMENT_REFERENCE_REUSED
            - STOCK_UNAVAILABLE
            - VALIDATION_FAILED
            - DOMAIN_ERROR
            - UPSTREAM_TIMEOUT
        correlationId:
          allOf:
            - $ref: '#/components/schemas/UniqueId'
          description: The id this failure was logged under (R58).
        occurredAt:
          $ref: '#/components/schemas/Instant'
      required: [title, status, code]
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                required: [field, message]
...